+++ packages/components/avatar/register.ts
@@ -1,9 +1,3 @@
 import { Avatar } from './src/avatar.js';
 
 customElements.define('sl-avatar', Avatar);
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'sl-avatar': Avatar;
-  }
-}