+++ packages/components/menu/src/menu-button.ts
@@ -6,12 +6,19 @@ import { type PopoverPosition } from '@sl-design-system/shared';
 import { type CSSResultGroup, LitElement, type PropertyValues, type TemplateResult, html, nothing } from 'lit';
 import { property, query, state } from 'lit/decorators.js';
 import styles from './menu-button.scss.js';
-import { MenuItem } from './menu-item.js';
 import { Menu } from './menu.js';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'sl-menu-button': MenuButton;
+  }
+}
 
 /**
  * Custom element that combines a button and a menu and automatically wires them up
  * together.
+ *
+ * @csspart button - The button element.
  *
  * @slot default - The menu items should be slotted in the default slot.
  * @slot button - Any content for the button should be slotted here.
@@ -83,6 +90,7 @@ export class MenuButton extends ScopedElementsMixin(LitElement) {
         .fill=${this.fill}
         .size=${this.size}
         .variant=${this.variant}
+        part="button"
       >
         <slot name="button"></slot>
         ${this.selects && this.selected ? html`<span class="selected">${this.selected}</span>` : nothing}
@@ -119,10 +127,8 @@ export class MenuButton extends ScopedElementsMixin(LitElement) {
     }
   }
 
-  #onMenuClick(event: Event): void {
-    if (event.target instanceof MenuItem) {
-      this.menu.hidePopover();
-    }
+  #onMenuClick(): void {
+    this.menu.hidePopover();
   }
 
   #onSelect(): void {