+++ packages/components/menu/src/menu-item.ts
@@ -7,13 +7,22 @@ import {
   event,
   isPopoverOpen
 } from '@sl-design-system/shared';
+import { SlSelectEvent } from '@sl-design-system/shared/events.js';
 import { type CSSResultGroup, LitElement, type PropertyValues, type TemplateResult, html, nothing } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import styles from './menu-item.scss.js';
 import { Menu } from './menu.js';
 
+declare global {
+  interface HTMLElementTagNameMap {
+    'sl-menu-item': MenuItem;
+  }
+}
+
 /**
  * Menu item component for use inside a menu.
+ *
+ * @csspart wrapper - The wrapper around the menu item content.
  *
  * @slot default - Content to display inside the menu item.
  * @slot submenu - The menu items that will be displayed when the menu item is shown.
@@ -48,7 +57,7 @@ export class MenuItem extends ScopedElementsMixin(LitElement) {
   @property({ type: Boolean, reflect: true }) disabled?: boolean;
 
   /** Emits when the user toggles the selected state. */
-  @event({ name: 'sl-select' }) selectEvent!: EventEmitter<boolean>;
+  @event({ name: 'sl-select' }) selectEvent!: EventEmitter<SlSelectEvent>;
 
   /** Whether this menu item has been selected. */
   @property({ type: Boolean, reflect: true }) selected?: boolean;
@@ -88,7 +97,7 @@ export class MenuItem extends ScopedElementsMixin(LitElement) {
   override render(): TemplateResult {
     return html`
       <div aria-hidden="true" class="safe-triangle"></div>
-      <div class="wrapper">
+      <div part="wrapper">
         ${this.selected ? html`<sl-icon name="check"></sl-icon>` : nothing}
         <slot></slot>
         ${this.shortcut ? html`<kbd>${this.#shortcut.render(this.shortcut)}</kbd>` : nothing}