+++ packages/components/radio-group/register.ts
@@ -3,10 +3,3 @@ import { Radio } from './src/radio.js';
 
 customElements.define('sl-radio', Radio);
 customElements.define('sl-radio-group', RadioGroup);
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'sl-radio': Radio;
-    'sl-radio-group': RadioGroup;
-  }
-}