+++ packages/components/shared/src/events/change.ts
@@ -1,0 +1,7 @@
+export type SlChangeEvent<T = unknown> = CustomEvent<T>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'sl-change': SlChangeEvent;
+  }
+}