+++ packages/components/spinner/src/spinner.ts
@@ -2,6 +2,12 @@ import { type CSSResultGroup, LitElement, type TemplateResult, html, type nothin
 import { property } from 'lit/decorators.js';
 import styles from './spinner.scss.js';
 
+declare global {
+  interface HTMLElementTagNameMap {
+    'sl-spinner': Spinner;
+  }
+}
+
 export type SpinnerSize = 'sm' | 'md' | 'lg' | 'xl' | '2xl' | '3xl';
 export type SpinnerVariant = 'accent' | 'info' | 'danger' | 'success' | 'warning';
 
@@ -12,13 +18,13 @@ export type SpinnerVariant = 'accent' | 'info' | 'danger' | 'success' | 'warning
  * <sl-spinner></sl-spinner>
  * ```
  *
- * @cssprop --sl-spinner-size - The size of the spinner, defaults to md if not set
+ * @cssprop --sl-spinner-size - The size of the spinner, defaults to `md` if not set.
  */
 export class Spinner extends LitElement {
   /** @private */
   static override styles: CSSResultGroup = styles;
 
-  /** The size of the spinner. Defaults to md in CSS if not set. */
+  /** The size of the spinner. Defaults to `md` with css properties if not attribute is not set. */
   @property({ reflect: true }) size?: SpinnerSize;
 
   /** The spinner variant. */
@@ -28,7 +34,8 @@ export class Spinner extends LitElement {
     super.connectedCallback();
 
     if (!this.hasAttribute('role')) {
-      this.setAttribute('role', 'alert');
+      this.setAttribute('role', 'presentation');
+      this.setAttribute('aria-hidden', 'true');
     }
   }