+++ packages/components/select/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sl-design-system/select
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [[`090c7b0`](https://github.com/sl-design-system/components/commit/090c7b039c8a7cadbdfbed0563764445d792c3da), [`090c7b0`](https://github.com/sl-design-system/components/commit/090c7b039c8a7cadbdfbed0563764445d792c3da)]:
+  - @sl-design-system/shared@0.2.8
+  - @sl-design-system/form@0.0.12
+
+## 0.0.24
+
+### Patch Changes
+
+- [#1076](https://github.com/sl-design-system/components/pull/1076) [`1958093`](https://github.com/sl-design-system/components/commit/19580930a666b18051aa0c64dafc7d9c0fdffaf4) - Fix incorrect color when selected & disabled
+
+- Updated dependencies [[`38b0ca4`](https://github.com/sl-design-system/components/commit/38b0ca4d72014605418639b69410863eb8e231ad)]:
+  - @sl-design-system/shared@0.2.7
+  - @sl-design-system/form@0.0.11
+
 ## 0.0.23
 
 ### Patch Changes